Initial seed data populating table

+++ migrations/20171208125014_assassins.js
@@ -8,7 +8,7 @@ exports.up = function(knex, Promise) {
     table.string('codeName', 50);
     table.string('weapon', 50).notNullable().defaultTo('None needed');
     table.string('contactInfo', 75).notNullable().defaultTo('Contact by fax');
-    table.integer('rating').notNullable().defaultTo(0);
+    table.float('rating').notNullable().defaultTo(0);
     table.integer('kills').notNullable().defaultTo(0);
     table.integer('price').notNullable().defaultTo(0);
     table.integer('age');
